fixed offsets to use longs instead of ints. JCT-117
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@776 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src/org/jcvi/common/core/assembly/contig/ace/IndexedAceFileDataStore.java
@@ -201,9 +201,9 @@ public final class IndexedAceFileDataStore implements AceContigDataStore{
     private static class IndexedAceFileDataStoreBuilder implements AceContigDataStoreBuilder{
         private IndexedFileRange indexFileRange;
         private final File aceFile;
-        private int currentStartOffset;
+        private long currentStartOffset;
         private int currentLineLength;
-        private int currentFileOffset;
+        private long currentFileOffset;
         private boolean firstContig=true;
         private String currentContigId;
